Split the statistics engine from the rest of the stats.
The engine is a singleton, and should be managed as such. Splitting it
out helps ensure that multiple agents don't end up inadvertently leaking
themselves into the event-handler queue for the transactions collection
(when they add themselves as listeners) through the stats engine.

Also added the ability to reset the statistics engine for testing.