Removed dependency on stub agent.
Because the Agent class has been split from the agent bootstrap module,
it's now possible to use it directly, with a little bit of fakery, to
test the connection to the collectors more natively. Ideally,
CollectorConnection wouldn't be dependent upon the agent for anything
other than the harvest cycle, so we could test the connection class
completely on its own, but until then, this method uses a lot less ganky
testing code to achieve the same result as before.