Reworked ApdexStats for clarity, and now test it.
Did a lot of the same cleanup as I did for the Stats object, with a
bunch of hedging around how ApdexStats serialization works (don't like
it, will probably fix it). Also wrote a bunch of tests and added a
couple functions to make it look and behave more like the Stats class.