Cut over to Metrics. Removed StatsEngine, Collection and old metric API.
A major architectural change that still isn't completely settled (some
tests will fail until the cut over to the new transaction tracing
system). With a little more work, the coupling between metrics,
instrumentation and the collector collection should be much looser.

As part of this work, wrote a bunch more tests, but I suspect a lot of
the instrumentation is actually broken and untested at this point.