Remove hacky binding of transaction to agent.
For better and for worse, attaching a transaction directly to the agent
isn't reliable, and having it there leads to a false sense of security.
At some point, we just have to trust the shimmer and callstack modules
to do their jobs. I think the agent has reached that point.