Sometimes it's the simple things that hurt the most.
Trying to ensure that HTTP requests to external services that are part
of an inbound request / transaction led me on a transcendent voyage of
discovery and pain. It took me 8 hours to realize that my code was doing
exactly what it was supposed to be doing, and another hour to add the
necessary code to actually test what I thought I was supposed to be
testing.

Kindly note the crazy nesting of web servers here, which actually
mirrors, in an inside-out way, how the requests to them are supposed to
work.