Nothing is ever simple at 2 a.m.
Added another test to ensure external request metrics were being merged
back into the agent's metrics after the transaction was finished and
discovered there was a chance that the metrics might be merged from the
transaction into the agent until all of the participants in the
transaction had finished reporting yet.

Added a simple reference-counting scheme so probes would have an
opportunity to record their metrics. In a multi-threaded world, there
would need to be a lot more synchronization to prevent race conditions,
but once again, Node keeps things simple and confusing.