Added methods to send more trace data to the collector.
Mainly this was to get transaction trace submission working, but I also
went ahead and put in the necessary support for the (as-yet unsupported)
slow SQL data. This entailed a bunch of work to come up with realistic
tests, and also some small modifications to the data sender, which needs
further work.