Agent now manages the update of metric renaming and normalization rules.
Some serious code envy going on here; was leading to an odd issue where
metric renaming rules would get pushed from the server, the IDs would be
pushed the next harvest cycle, and then would revert to the metric specs
on the the harvest after. It doesn't make sense for an object that's
getting recreated periodically to try to hold onto this state anyway.

That said, the agent class is starting to feel like a giant grab bag of
stuff again; probably needs to be factored into more tightly-coupled
pieces again.