Deal with request parameters.
Attaches them to both the trace segment and the root of the trace, and
strips them properly from request URLs and scopes. Assumes that we want
to know when parameters are present but not assigned a value and sends a
bare JSON true value to the collector when these are encounted. Could be
wrong! Only time will tell!