Allow measurements to be added to transaction after timer has finished.
In some sense the guard on allowing measurements after transaction's end
conflicts with the cheap observer pattern implemented by pending
transaction reporters. The transaction isn't finished until the
reporters are finished, which is risky in that instrumentation that
never finishes will just leave the transaction hanging in the breeze.
Solution is to not write any leaky instrumentation?