'k1&k2=v' is not necessarily the same as 'k1=&k2=v'.
Add a check to ensure that if a variable name is a bareword (i.e.
there's no trailing '='), its presence is noted by setting its
value to true (to indicate that it was present). Empty variables will
still have an empty value like usual.

As the comment says, this might just be more confusing than it's worth
for everybody, so it might not last.