More or less rewrote CollectorConnection to be simpler.
The old approach used a tough-to-understand mixture of normal,
callback-based Node programming and semi-dynamic overrides via
passed-in callbacks and errorbacks. The new approach is almost
entirely evented, which has the bonus side effect of making the
connection far more observable -- see the additional tests to
the connection integration test case to see how observable it is
now. Every step in the connection's life cycle now has an
associated event.

A lot of work was put into improving the logging, as well.