Don't break functions that check callback arity.
Given the way that many JavaScript functions reflect on the arguments object
to make themselves variadic, doing quasi-dynamic dispatch on the length of
the declared arguments list is a bad idea for a host of reasons. That said,
Connect, Express and Mocha all do it. The only way to set arity is by
construction a function expression with the appropriate declarations, so it's
unfortunately necessary to add another function call onto the stack to avoid
breaking things.