Pull versioned tests up a level, make more tests versioned.
Created two new versioned instrumentaiton tests, for node-mongodb-native
(which has a very brittle internal structure and which causes the
instrumentation to break across minor versions), and for node-mysql-0-9.
The node-mongodb-native tests are transitional -- going to replace them
with a multiverse test suite that will ensure that whatever solution we
come up with for the instrumentation will be stable across many versions
of the driver.

WIP: need to ensure that all of the correct dependencies are being
installed via the Makefile -- and that they're not being reinstalled
every single time the integration tests are run. Also need to make sure
that architect infrastructure has its dependencies installed as locally
as possible.