Run tests for each multiverse directory.
There's probably a way to do this without using yet another control flow
module, but I was too dumb tonight to find the correct way to do this
with promises, seeing as how I want to sequence both the installation of
the various versions of the modules *and* run the tests. The important
part is that I have a highline summary of how many tests are failing per
module version.