metrics: tie apdex recording to ignored statuses
The agent already ignores certain status codes for the purposes of error
tracing, and it seems perfectly consistent to not automatically mark any
ignored status codes as frustrating for Apdex purposes.

Also started refactoring measureWeb into smaller pieces. It'll probably
take a few steps, as there's a lot of gross mutation happening in that
function.