instrumentation: extracted metrics recorders
Now there is a set of simple modules that record metrics, allowing a set
of standard tests to be built against them, and combining several of
them into a generic, simple metrics recorder. Some of the existing
instrumentation was inconsistent in how it captured metrics, so I
rewrote those to clean that up. It had been a while since the fs.readdir
instrumentation had gotten any love, so it got some dramatic cleanup, as
did its integration test.

The lack of test failures while I was doing this work, sometimes around
code that was flat-out incorrect, leads me to believe there's lots of
room for better test coverage.