code: prevent deoptimization
V8 will create a hidden class when optimizing code, and if any
properties are access that are not on this hidden class, it will trigger
a deoptimization. Simply creating the field and initializing it to null
is enough to address this. JavaScript works better when you pretend
it's statically typed.