errors: merge correctly on failed collector run
Before, the error tracer was trying to merge the errors object passed to
the DataSender, which is to say the raw JSON expected by the collector.
This was then getting redelivered to the collector, causing it to barf
when unmarhsalling the error traces.

The fix also involved moving responsibility for merging the data into
the agent from the error tracer, and renaming some methods to more
clearly reflect what they were doing. This responsibility should
probably move to the collector connection, so as to keep everything that
deals with collector data formats in one place.