config: disable capture_params, add ignored_params
So this adds support for SSC of capturing params off requests for
transaction traces and error traces, and it also adds support (which
needs to be wired up on the server) for ignoring request parameters as
well. Both are now documented in the default configuration file, as well
as having corresponding environment variable configurations.