instrumentation: filter self-calls from traces
See the code comments in the commit for details, but to give a single
example, collection.findAndRemove calls colleciton.findAndModify, and
it's highly unlikely developers know or care about that implementation
detail. Figuring out what's going on requires a little trickery with CLS
to keep track of the types of segments created on the current tick,
which is a strong enough filter to tell the difference between nested
and immediate self-calls.

Also symlinked the multiverse MongoDB tests into the specific versioned
tests. They're too comprehensive to leave sitting on their own.