test: split MongoDB tests into separate versions
Doing it as a symbolic link to the multiverse tests was DRY, but then it
was using the multiverse's deps, which was at best confusing and at
worst totally incorrect. The tests are actually only completely passing
with the latest version of node-mongodb-native (1.3.19), and the fact
that they were passing with 1.1.7 was a fluke based on what version was
installed in multiverse.