collector: put response parsing in one function
...and then put that function in its own file, and give it its own
tests. The response parser is as resilient as I could make it, and
unifies errors on the response stream, JSON parsing errors, exceptions
sent by the server, and non-OK HTTP status codes. It also tries to
extract data and pass it along whenever possible, even when there are
errors. This allows remote method invocation and the collector API to be
more flexible, as well as ensuring that errors all have certain
consistent properties.

Accumulating all the errors and passing them along allows the
higher-level API to provide a complete transcript of what went wrong
with a particular set of calls to the collector, which should make the
logs much more informative to support and the dev team. It also
simplifies the code a little more, which is a nice bonus.