collector: fill out public API
I've gotten a little head of my tests and will have to catch up, but
there are tests for the most important new chunk, which is
_runLifecycle, which will call a method and then deal with common
responses, including discarding bogus requests, reconnecting on
ForceRestartException, and shutting down on ForceDisconnectException.

I added a small optimization, which is that there's only one instance
of each remote method. This should save a tiny bit of GC, but it was
easy to do, so I just went for it.