agent: responsibile for serializing data, not API
Also added documents to all the methods in the API and the agent.

Figuring out which piece of the module is responsible for which
functions has bedeviled me for most of the last two years. The agent
acts as a mediator between the collector API and the rest of the agent,
and the API should map as closely to the collector's view of the world
as possible, which implies that it's in charge of marshalling data and
handling the responses sent by the collector, not being responsible for
formatting that data.

At the same time, in order for the API to be able to seamlessly
reconnect to the collector, it needs to be able to update the module's
configuration. The API's view of configuration is read-only, and to
preserve that decoupling, the agent needs to expose a method the API can
use to request an update to the configuration.

So I had to move a bunch of logic between the API and the agent, and the
result complicates the logic in the agent, but everything that is in
there now is where it belongs, and the API is much simpler as a result.