remote-method: Remove unnecssary DNS lookup
Back in November 2013, Forrest added a change to perform a preemptive
DNS lookup of the collector hostname before attempting to connect to it
because he found that http.request() was throwing an error on lookup
failures. I found a commit in Node.js core that would cause this to
occur:

https://github.com/joyent/node/commit/ca9eb718fbf2cd2c60c7aeb3ca33413e17fcbbf0

This was released as part of v0.11.5 on July 18th, 2013. Forrest, being
a dutiful Node.js developer, was probably using one of these latest
unstable releases when he found this issue. Thankfully, a fix was
committed recently:

https://github.com/joyent/node/commit/61baefce1e8cae13e40f01c04c8f684848cee3e2

Unfortunately, this fix has not been released yet. That said, we don't
support unstable releases of Node.js, partly so we don't need to add in
ugly workarounds for brokenness such as this.

This change removes the unnecessary lookup, and resolves issues where
customers behind a proxy could not resolve the collector hostname.