pg: Fix for using query events instead of callbacks
When no callback was detected, we were inserting our own. The insertion
itself caused a crash, but on top of that adding a callback modifies the
behavior of the pg module. Instead, we now listen for error or end
events to finish segments.

We also now catch new segments started in event handlers properly.