Pass through additional arguments to memcached.command .
For some calls -- in particular, getMulti -- memcached.command takes an
extra argument that forces the choice of a particular server.

The memcached instrumentation was eating this extra parameter and not
passing it on to the wrapped implementation, which caused the library to
consult only one memcached server for getMulti, giving the appearance of
a cache miss for anything not on that server.

Instead we now pass all of the additional arguments through verbatim,
fixing the multi-server behavior and generally being a little more polite
to the interface of the underlying memcached library.