Don't add interceptor (error middleware) to Routes stack
The current problem is that the error middleware that new relic adds to the main stack is also being added to every Router method's stack.
This means that the error middleware is actually being run twice, once from the Router method's stack and once from the general stack.
Besides this being wrong the other problem is that the cleanup function (line 398) that runs after the first error middleware finished removes the partialName thus causing the errors in new relic to appear without the specific url. it will only show '/' in the Errors view because of the cleanup.
The solution is to not add the error middleware for Router stacks. Add it only for the main stack.