Undefined check on segment transaction trace
We're getting some errors when doing a JSON.stringify:

```
TypeError: Cannot read property 'root' of undefined
    at TraceSegment.toJSON
(/src/node_modules/newrelic/lib/transaction/trace/segment.js:389:75)
    at JSON.stringify (<anonymous>)
    at <our code...>
```

It appears that `segment.transaction.trace` is `undefined`. Note that
this occurs during a connection timeout, in case you want to fix the
underlying bug where that is not populated. I'm not sure where the trace
segment is getting attached to the object tree we're serializing, but I
thought you'd probably want this `TypeError` fixed in any case.