facts: add supportability for some popular logging libraries
I am not quite sure but I think the environment's packages seems to be
the right place to look for these. Those are logging libraries
directly mentioned in packages.json, as far as I can tell. This is as
opposed to the environment's dependencies which seems to be all the
packages the current packages depend on.