grpc: remove host/port from server-side instrumentation
Although the "spec" says that `request.uri` should be
`grpc://<host>:<port>/<request method>`, we don't actually report host
and port for any other server-side instrumentation. It's complicated
to get that information from the channelz property to begin with, so
best to just not even dig around a deeply nested property to begin
with.