NR-69739 add batching to dep lookup
In 9.0.1, we removed the async library as a dependency because as a group we did not like how confusing it was to use. An unfortunately side effect of that is that we removed throttling the dependency lookups at start-up. This meant that customers could run into EMFILE issues, as shown in the GTSE attached to the ticket