feat: add ability to customize Error Groups
New API function called `setErrorGroupCallback` provides a way for you
to customize the `error.group.name` attribute of errors that are
captured by the agent. This attribute controls how the Errors Inbox
functionality groups similar errors together.

`setErrorGroupCallback` accepts one argument: a callback function. This
function can accept an object argument, which contains metadata
associated with an individual error, and must return a string, which
will be used as the Error Group name. The object passed to the callback
function has the following shape:
```
{
  error: Error,
  customAttributes: Object,
  'request.uri': String,
  'http.statusCode': String,
  'http.method': String,
  'error.expected': Boolean
}
```