ci: script to generate release notes from commits
Created the ConventionalChangelog class, which has methods for parsing
conventional commits from the git log, as well as methods for generating
JSON and Markdown changelogs. JSON changelogs will be used by the
docs-website PR script to generate the front-matter for Agent Version
Metadata service. Markdown changelogs are for our own documentation and
for customers to read