chore: pin testdobule to 3.17.2 as 3.18.0 no longer works on node 14 (#1628)
<!--
Thank you for submitting a Pull Request.

This code is leveraged to monitor critical services. Please consider the
following:
* Tests are required.
* Performance matters.
* Features that are specific to just your app are unlikely to make it
in.

Please fill out the relevant sections as follows:
* Proposed Release Notes: Bulleted list of recommended release notes for
the change(s).
* Links: Any relevant links for the change.
* Details: In-depth description of changes, other technical notes, etc.
-->

## Proposed Release Notes

## Links

## Details
We have to pin `testdouble` until we drop support for node 14.
[3.18.0](https://github.com/testdouble/testdouble.js/blob/main/CHANGELOG.md#3180)
release added support for Node 20 which even though it doesn't call it
out dropped support for 14 and it's breaking our esm unit tests.