fix: disable running CI for releases
There are a few parts to this:

1) First, disable running on pushes. This is because pushes don't have
context when they come from a pull request. In particular, pushes
don't have the branch name the pull request came from.

2) Then, disable running if the branch name starts with release.

3) In the same condition check, enable runs only for branches that are
closed via a merge. We don't want to trigger CI for a PR that is
closed because we decided to abandon the idea.