refactor getVendors to an async function
The sole usage of the function within the agent was already promisifying
the function. So it was already being used as an async function.
By refactoring it to an actual async function, we can process vendors
serially. This will make it easier to review trace logs as vendor
resolution messages will no longer be interleaved.